Commit 18/09/2024; Updated Invoicing system capturing all correct details depending on Product Category registered via Cycle. Sales Form is also generated dynamically as per the details above, i.e Product Category. Modified Models and tables to cater for such updates. Created seeders with demo data for Customers, Customer Salesperson. Created partials for generating forms for sales depending on the case.

+++ database/migrations/2024_03_09_175253_create_customers_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('customers', function (Blueprint $table) {
             $table->id();
             $table->string('Customer_Name', 255);
-            $table->integer('Cust_Account_No');
+            $table->bigInteger('Cust_Account_No');
             $table->string('Address', 255);
             $table->timestamps();
